Session 2/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-genai==0.2.0
beautifulsoup4==4.12.2
geopy==2.4.1
pillow==10.1.0
